general fixing
I changed the design of the product page, also changed the navbar,
category list.
I fixed some issues on the register page.
Fixed the home page on the admin side.
Changed the layout on view products on the admin side.
Used the imageProducts class on the customer side to make sure it
displays the images of the products.